dvfs: Mv bdev_by_path() to devfs
