+++ src/fs/driver/devfs/devfs_dvfs.c
@@ -28,8 +28,6 @@
 
 #include <util/array.h>
 
-#include <module/embox/driver/block_dev.h>
-
 extern struct idesc_ops idesc_bdev_ops;
 extern int devfs_destroy_inode(struct inode *inode);
 
@@ -80,3 +78,41 @@ static const struct fs_driver devfs_dumb_driver = {
 DECLARE_FILE_SYSTEM_DRIVER(devfs_dumb_driver);
 
 FILE_SYSTEM_AUTOMOUNT("/dev", devfs_dumb_driver);
+
+struct block_dev *bdev_by_path(const char *dev_name) {
+	struct lookup lookup = {};
+	struct dev_module *devmod;
+	int res;
+
+	if (!dev_name) {
+		return NULL;
+	}
+	if (!strlen(dev_name)) {
+		return NULL;
+	}
+
+	/* Check if devfs is initialized */
+	res = dvfs_lookup("/dev", &lookup);
+	if (res) {
+		/* devfs is not mounted yet */
+		return block_dev_find(dev_name);
+	}
+	dentry_ref_dec(lookup.item);
+
+	/* devfs presents, perform usual mount */
+	memset(&lookup, 0, sizeof(lookup));
+	dvfs_lookup(dev_name, &lookup);
+	if (!lookup.item) {
+		SET_ERRNO(ENOENT);
+		return NULL;
+	}
+
+	assert(lookup.item->d_inode);
+
+	devmod = inode_priv(lookup.item->d_inode);
+
+	dentry_ref_dec(lookup.item);
+
+	return dev_module_to_bdev(devmod);
+}
+
